feat: about and team page completed
