+++ src/app/app.tsx
@@ -13,7 +13,7 @@ const App: Layout = ({ children }) => {
   return (
     <div className="min-h-screen bg-slate-900">
       <Header />
-      <main id="main">{children}</main>
+      {children}
       <Footer />
     </div>
   );
