+++ src/components/AnimatedText.tsx
