+++ src/data/about/coreValues.ts
@@ -1,0 +1,26 @@
+import { Heart, Shield, Target, Users } from 'lucide-react';
+
+const coreValues = [
+  {
+    icon: Heart,
+    title: 'Inclusion',
+    description: 'We welcome developers of all skill levels and backgrounds',
+  },
+  {
+    icon: Users,
+    title: 'Collaboration',
+    description: 'We believe in learning together and supporting each other',
+  },
+  {
+    icon: Target,
+    title: 'Growth',
+    description: 'We focus on continuous learning and skill development',
+  },
+  {
+    icon: Shield,
+    title: 'Respect',
+    description: 'We maintain a safe and respectful environment for all',
+  },
+];
+
+export default coreValues;
