+++ src/data/team/teamMembers.ts
@@ -1,0 +1,124 @@
+const teamMembers = [
+  {
+    name: 'Rajesh Patel',
+    role: 'Community Lead',
+    image: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events',
+    slug: 'rajesh-patel',
+    category: 'Core',
+  },
+  {
+    name: 'Rajesh Patel',
+    role: 'Community Lead',
+    image: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events',
+    slug: 'rajesh-patel',
+    category: 'Core',
+  },
+  {
+    name: 'Priya Shah',
+    role: 'Event Coordinator',
+    image: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'priya-shah',
+    category: 'Core',
+  },
+  {
+    name: 'Priya Shah',
+    role: 'Event Coordinator',
+    image: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'priya-shah',
+    category: 'Core',
+  },
+  {
+    name: 'Priya Shah',
+    role: 'Event Coordinator',
+    image: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'priya-shah',
+    category: 'Core',
+  },
+  {
+    name: 'Amit Kumar',
+    role: 'Technical Mentor',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'amit-kumar',
+    category: 'Mentor',
+  },
+  {
+    name: 'Amit Kumar',
+    role: 'Technical Mentor',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'amit-kumar',
+    category: 'Mentor',
+  },
+  {
+    name: 'Amit Kumar',
+    role: 'Technical Mentor',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'amit-kumar',
+    category: 'Mentor',
+  },
+  {
+    name: 'Amit Kumar',
+    role: 'Technical Mentor',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'amit-kumar',
+    category: 'Mentor',
+  },
+  {
+    name: 'Amit Kumar',
+    role: 'Technical Mentor',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Passionate about building tech communities and organizing impactful events.',
+    slug: 'amit-kumar',
+    category: 'Mentor',
+  },
+  {
+    name: 'John Volunteer',
+    role: 'Volunteer',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Helps organize and run community events.',
+    slug: 'john-volunteer',
+    category: 'Volunteer',
+  },
+  {
+    name: 'John Volunteer',
+    role: 'Volunteer',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Helps organize and run community events.',
+    slug: 'john-volunteer',
+    category: 'Volunteer',
+  },
+  {
+    name: 'John Volunteer',
+    role: 'Volunteer',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Helps organize and run community events.',
+    slug: 'john-volunteer',
+    category: 'Volunteer',
+  },
+  {
+    name: 'John Volunteer',
+    role: 'Volunteer',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Helps organize and run community events.',
+    slug: 'john-volunteer',
+    category: 'Volunteer',
+  },
+  {
+    name: 'John Volunteer',
+    role: 'Volunteer',
+    image: 'https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?w=400',
+    bio: 'Helps organize and run community events.',
+    slug: 'john-volunteer',
+    category: 'Volunteer',
+  },
+];
+
+export default teamMembers;
